Ending tracking of dist, moving resources to respective places in src.
