+++ src/images/thumbnails/Spotify_Icon_RGB_Green.png
