+++ src/styles/style.css
@@ -9,7 +9,7 @@ body {
     width: 100%;
     margin: 0;
     padding: 0;
-    background: url("../images/painting.jpg");
+    background: url("painting-bg.jpg");
     background-size: cover;
     font-family: monospace;
